scimulator/synthetic_demand_engine/requirements.txt
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
pyyaml>=6.0
pytest>=7.4.0
